Work on generator

+++ hapi-tinder-plugin/src/test/java/ca/uhn/fhir/tinder/ddl/test/ComplexId.java
@@ -3,6 +3,6 @@ package ca.uhn.fhir.tinder.ddl.test;
 public class ComplexId {
 
 	private Long myId;
-	private Double myPartitionId;
+	private Integer myPartitionId;
 
 }
